TINKERPOP-3010 Move TinkerTransactionGraph-specific tests.
